Collapse store and resource handler method argument resolvers
