--- spring-content-rest/src/main/java/internal/org/springframework/content/rest/controllers/StoreHandlerMethodArgumentResolver.java
@@ -1,178 +1,0 @@
-package internal.org.springframework.content.rest.controllers;
-
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.springframework.content.commons.mappingcontext.ContentProperty;
-import org.springframework.content.commons.mappingcontext.MappingContext;
-import org.springframework.content.commons.property.PropertyPath;
-import org.springframework.content.commons.repository.AssociativeStore;
-import org.springframework.content.commons.repository.Store;
-import org.springframework.content.commons.storeservice.StoreInfo;
-import org.springframework.content.commons.storeservice.Stores;
-import org.springframework.content.rest.config.RestConfiguration;
-import org.springframework.context.ApplicationContext;
-import org.springframework.core.MethodParameter;
-import org.springframework.core.io.Resource;
-import org.springframework.data.repository.support.Repositories;
-import org.springframework.util.AntPathMatcher;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-import org.springframework.web.util.UriTemplate;
-import org.springframework.web.util.UrlPathHelper;
-
-import internal.org.springframework.content.rest.controllers.resolvers.EntityResolution;
-import internal.org.springframework.content.rest.controllers.resolvers.EntityResolvers;
-import internal.org.springframework.content.rest.io.AssociatedStorePropertyPathResourceImpl;
-import internal.org.springframework.content.rest.utils.StoreUtils;
-
-public abstract class StoreHandlerMethodArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private UriTemplate entityUriTemplate = new UriTemplate("/{repository}/{id}");
-    private UriTemplate entityPropertyUriTemplate = new UriTemplate("/{repository}/{id}/{property}");
-    private UriTemplate entityPropertyWithIdUriTemplate = new UriTemplate("/{repository}/{id}/{property}/**");
-    private UriTemplate revisionPropertyUriTemplate = new UriTemplate("/{repository}/{id}/revisions/{revisionId}/{property}");
-    private UriTemplate revisionPropertyWithIdUriTemplate = new UriTemplate("/{repository}/{id}/revisions/{revisionId}/{property}/{contentId}");
-
-    private EntityResolvers entityResolvers;
-    private List<ResourceResolver> resolvers = new ArrayList<>();
-
-    private ApplicationContext context;
-    private final RestConfiguration config;
-    private final Repositories repositories;
-    private final Stores stores;
-    private final MappingContext mappingContext;
-
-    public StoreHandlerMethodArgumentResolver(ApplicationContext context, RestConfiguration config, Repositories repositories, Stores stores, MappingContext mappingContext, EntityResolvers entityResolvers) {
-        this.context = context;
-        this.config = config;
-        this.repositories = repositories;
-        this.stores = stores;
-        this.mappingContext = mappingContext;
-
-        this.entityResolvers = entityResolvers;
-
-        resolvers.add(new StoreResourceResolver());
-        resolvers.add(new AssociativeStoreResourceResolver());
-    }
-
-    RestConfiguration getConfig() {
-        return config;
-    }
-
-    protected Repositories getRepositories() {
-        return repositories;
-    }
-
-    protected Stores getStores() {
-        return stores;
-    }
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return Store.class.isAssignableFrom(parameter.getParameterType());
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-
-        String pathInfo = webRequest.getNativeRequest(HttpServletRequest.class).getRequestURI();
-        pathInfo = new UrlPathHelper().getPathWithinApplication(webRequest.getNativeRequest(HttpServletRequest.class));
-        pathInfo = StoreUtils.storeLookupPath(pathInfo, this.getConfig().getBaseUri());
-
-        String[] pathSegments = pathInfo.split("/");
-        if (pathSegments.length < 2) {
-            return null;
-        }
-
-        String store = pathSegments[1];
-
-        StoreInfo info = this.getStores().getStore(Store.class, StoreUtils.withStorePath(store));
-        if (info == null) {
-            throw new IllegalArgumentException(String.format("Store for path %s not found", store));
-        }
-
-        if (AssociativeStore.class.isAssignableFrom(info.getInterface())) {
-
-            EntityResolution result = this.entityResolvers.resolve(pathInfo);
-
-            AntPathMatcher matcher = new AntPathMatcher();
-            Comparator<String> patternComparator = matcher.getPatternComparator(pathInfo);
-
-            List<String> uriTemplates = new ArrayList<>();
-            for (ResourceResolver resolver : resolvers) {
-                if (matcher.match(resolver.getMapping(), pathInfo)) {
-                    uriTemplates.add(resolver.getMapping());
-                }
-            }
-
-            String bestMatch = null;
-            if (uriTemplates.size() > 1) {
-                uriTemplates.sort(patternComparator);
-            }
-
-            bestMatch = uriTemplates.get(0);
-
-            ResourceResolver matchedResolver = null;
-            for (ResourceResolver resolver : resolvers) {
-                if (bestMatch.equals(resolver.getMapping())) {
-                    matchedResolver = resolver;
-                }
-            }
-
-            return matchedResolver.resolve(webRequest, info, result.getEntity(), result.getContentProperty());
-
-        } else if (Store.class.isAssignableFrom(info.getInterface())) {
-
-            return resolveStoreArgument(webRequest, info);
-        }
-
-        throw new IllegalArgumentException();
-    }
-
-    protected abstract Object resolveStoreArgument(NativeWebRequest nativeWebRequest, StoreInfo info);
-
-    protected abstract Object resolveAssociativeStoreEntityArgument(StoreInfo info, Object entity);
-
-    protected abstract Object resolveAssociativeStorePropertyArgument(StoreInfo storeInfo, Object domainObj, Object propertyVal, boolean embeddedProperty);
-
-    private static interface ResourceResolver {
-        public String getMapping();
-        public Resource resolve(NativeWebRequest nativeWebRequest, StoreInfo info, Object domainObj, ContentProperty property);
-    }
-
-    public static class StoreResourceResolver implements ResourceResolver {
-
-        @Override
-        public String getMapping() {
-            return "/{repository}/{id}";
-        }
-
-        @Override
-        public Resource resolve(NativeWebRequest nativeWebRequest, StoreInfo info, Object domainObj, ContentProperty property) {
-//            Resource r = info.getImplementation(AssociativeStore.class).getResource(domainObj);
-//            return new AssociatedStoreResourceImpl(info, domainObj, r);
-            Resource r = info.getImplementation(AssociativeStore.class).getResource(domainObj, PropertyPath.from(property.getContentPropertyPath()));
-            return new AssociatedStorePropertyPathResourceImpl(info, property, domainObj, r);
-        }
-    }
-
-    public static class AssociativeStoreResourceResolver implements ResourceResolver {
-
-        @Override
-        public String getMapping() {
-            return "/{repository}/{id}/**";
-        }
-
-        @Override
-        public Resource resolve(NativeWebRequest nativeWebRequest, StoreInfo info, Object domainObj, ContentProperty property) {
-            Resource r = info.getImplementation(AssociativeStore.class).getResource(domainObj, PropertyPath.from(property.getContentPropertyPath()));
-            return new AssociatedStorePropertyPathResourceImpl(info, property, domainObj, r);
-        }
-    }
-}
